requirements.txt
Flask==2.2.5
Werkzeug==2.2.3
Flask-SQLAlchemy==3.0.2
SQLAlchemy==1.4.41
psycopg2-binary==2.9.1
requests==2.26.0
numpy==1.20.3
pandas==1.2.5
python-dotenv==0.19.0
Flask-Migrate==3.1.0
Flask-Bootstrap4==4.0.2
plotly==5.3.1
PyJWT==2.6.0 